requirements.txt
torch==2.5.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn==1.5.1
kornia>=0.6.0
optuna>=3.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=6.0.0
tqdm>=4.62.0
pandas>=1.3.3
pillow>=8.3.2
tensorboard>=2.7.0
tqdm>=4.62.3
albumentations>=1.0.0
pytorch-lightning>=2.0.0
wandb>=0.12.0  # for experiment tracking
kaggle>=1.5.0  # for dataset download
pathlib>=1.0.1
shap>=0.39.0
lime>=0.2.0
captum>=0.4.0
torchviz>=0.0.2
graphviz>=0.17

# Deployment dependencies
onnx>=1.10.0
tf2onnx>=1.9.3
tensorflow==2.12.0
coremltools==7.1
fastapi>=0.70.0
uvicorn>=0.15.0
prometheus-client>=0.12.0
mlflow>=1.21.0
scipy>=1.7.3  # for drift detection
python-multipart>=0.0.5  # for FastAPI file uploads
